Texture is now a trait and the texture atlas and old texture (now StaticTexture) are merged in same file
